% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_AlecWebsite.R
\name{get_primary_color}
\alias{get_primary_color}
\title{Primary plotting color for Alec's Personal Website}
\usage{
get_primary_color()
}
\description{
Primary plotting color for Alec's Personal Website
}
